Add files via upload
This page is used by the user to view the doctor details along with his or her user reviews. They can also book an appointment on this page itself using the interface provided. THey can book one a week in advance.
Need to work on CSS for the doctor Profile pages.